+++ json2x.vcxproj
@@ -134,13 +134,13 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="src\jsonConverters.cpp" />
+    <ClCompile Include="src\CodeGenerator.cpp" />
     <ClCompile Include="src\json2x.cpp" />
     <ClCompile Include="src\win32.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\LangFormat.h" />
-    <ClInclude Include="src\jsonConverters.h" />
+    <ClInclude Include="src\CodeGenerator.h" />
     <ClInclude Include="src\win32.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
