+++ src/CodeGenerator.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include <string>
-#include "jsonConverters.h"
+#include "CodeGenerator.h"
 #include "rapidjson/error/en.h"
 
 #define FMT_HEADER_ONLY
