+++ src/json2x.cpp
@@ -2,9 +2,9 @@
 #include <fstream>
 #include <sstream>
 #include <vector>
-#include "jsonConverters.h"
-#include "win32.h"
 #include <csignal>
+#include "CodeGenerator.h"
+#include "win32.h"
 
 #define VERSION_MAJOR 1
 #define VERSION_MINOR 0
